Training Data Generation

To generate training data for the slowest non-linear equations found during Profiling Modelica Models we now simulate the equations multiple time and save in- and outputs.

We will use the Functional Mock-up Interface (FMI) standard to generate FMU that we extend with some function to evaluate single equations without the need to simulate the rest of the model.

Functions

NonLinearSystemNeuralNetworkFMU.generateTrainingData — Function

generateTrainingData(fmuPath, workDir, fname, eqId, inputVars, inputBounds, outputVars;
                     options=DataGenOptions())

Generate training data for given equation of FMU.

Generate random inputs between min and max, evalaute equation and compute output. All input-output pairs are saved in CSV file fname.

Arguments

fmuPath::String: Path to FMU.

workDir::String: Working directory for generateTrainingData.

fname::String: File name to save training data to.

eqId::Int64: Index of equation to generate training data for.

inputVars::Array{String}: Array with names of input variables.

inputBounds::MinMaxBoundaryValues{T}: Boundary values for input space.

outputVars::Array{String}: Array with names of output variables.

Keywords

options::DataGenOptions: Settings for data generation.

See also generateFMU, DataGenOptions..

NonLinearSystemNeuralNetworkFMU.addEqInterface2FMU — Function

addEqInterface2FMU(modelName, pathToFmu, eqIndices; workingDir=pwd())

Create extendedFMU with special_interface to evalaute single equations.

Arguments

modelName::String: Name of Modelica model to export as FMU.

pathToFmu::String: Path to FMU to extend.

eqIndices::Array{Int64}: Array with equation indices to add equiation interface for.

Keywords

workingDir::String=pwd(): Working directory. Defaults to current working directory.

Returns

Path to generated FMU workingDir/<modelName>.interface.fmu.

See also profiling, generateFMU, generateTrainingData.

NonLinearSystemNeuralNetworkFMU.generateFMU — Function

generateFMU(modelName, moFiles; options)

Generate 2.0 Model Exchange FMU for Modelica model using OMJulia.

Arguments

modelName::String: Name of the Modelica model.

moFiles::Array{String}: Path to the *.mo file(s) containing the model.

Keywords

options::OMOptions: Options for OpenModelica compiler.

Returns

Path to generated FMU workingDir/<modelName>.fmu.

See also OMOptions, addEqInterface2FMU, generateTrainingData.

Structures

NonLinearSystemNeuralNetworkFMU.DataGenOptions — Type

DataGenOptions <: Any

Settings for data generation.

method::NonLinearSystemNeuralNetworkFMU.DataGenerationMethod: Method to generate data points. Allowed values: RandomMethod, RandomWalkMethod

n::Integer: Number of data points to generate.

nBatches::Integer: Number of batches to divide N into.

nThreads::Integer: Number of threads to use in parallel

append::Bool: Append to already existing data

clean::Bool: Clean up temp CSV files

See als RandomMethod, RandomWalkMethod.

NonLinearSystemNeuralNetworkFMU.RandomMethod — Type

RandomMethod <: DataGenerationMethod

Random data generation using rand.

NonLinearSystemNeuralNetworkFMU.RandomWalkMethod — Type

RandomWalkMethod <: DataGenerationMethod

Randomized brownian-like motion data generation. Tries to stay within one solution in case the non-linear system is not unique solveable. Uses previous data point to generate close data point with previous solution as input to NLS solver.

delta::Float64: Step size of random walk.

Examples

First we need to create a Model-Exchange 2.0 FMU with OpenModelica.

This can be done directly from OpenModelica or with generateFMU:

moFiles = ["test/simpleLoop.mo"]
options = OMOptions(workingDir = "tempDir")

fmu = generateFMU("simpleLoop",
                  moFiles;
                  options = options)

"tempDir/simpleLoop.fmu"

Next we need to add non-standard FMI function

fmi2Status myfmi2EvaluateEq(fmi2Component c, const size_t eqNumber)

that will call <modelname>_eqFunction_<eqIndex>(DATA* data, threadData_t *threadData) for all non-linear equations we want to generate data for.

Using addEqInterface2FMU this C code will be generated and added to the FMU.

interfaceFmu = addEqInterface2FMU("simpleLoop",
                                  fmu,
                                  [14],
                                  workingDir = "tempDir")

"tempDir/simpleLoop.interface.fmu"

Now we can create evaluate equation 14 for random values and save the outputs to generate training data.

using CSV
using DataFrames
options = DataGenOptions(n=10, nThreads=1)
boundary = MinMaxBoundaryValues([0.0, 0.95], [1.5, 3.15])
generateTrainingData(interfaceFmu,
                     "tempDir",
                     "simpleLoop_data.csv",
                     14,
                     ["s", "r"],
                     boundary,
                     ["y"];
                     options = options)
df =  DataFrame(CSV.File("simpleLoop_data.csv"))

10×4 DataFrame

Row	s	r	y	Trace

Float64	Float64	Float64	Int64

1	0.306416	1.37214	1.15742	1

2	0.305057	1.37287	1.15732	1

3	0.305941	1.37348	1.1583	1

4	0.306241	1.37199	1.1572	1

5	0.306522	1.37187	1.15726	1

6	0.305455	1.37087	1.15584	1

7	0.306601	1.37015	1.15585	1

8	0.307054	1.37012	1.15606	1

9	0.30784	1.37025	1.15659	1

10	0.307048	1.36928	1.15535	1